rename models to xml_models
