+++ corehq/apps/app_manager/suite_xml/scheduler.py
@@ -2,7 +2,7 @@ from corehq.apps.app_manager.const import SCHEDULE_DATE_CASE_OPENED, SCHEDULE_LA
     SCHEDULE_GLOBAL_NEXT_VISIT_DATE, SCHEDULE_NEXT_DUE
 from corehq.apps.app_manager.exceptions import ScheduleError
 from corehq.apps.app_manager.suite_xml.const import FIELD_TYPE_SCHEDULE
-from corehq.apps.app_manager.suite_xml.models import DetailVariable
+from corehq.apps.app_manager.suite_xml.xml_models import DetailVariable
 from corehq.apps.app_manager.templatetags.xforms_extras import trans
 from corehq.apps.app_manager.xpath import ScheduleFormXPath
